+++ packages/axe-core-cucumber/e2e/capybara/features/support/env.rb
@@ -1,9 +1,9 @@
 require "capybara/cucumber"
 # require driver of choice
 require "axe-capybara"
-# Requiring 'axe/cucumber/step_definitions' makes all of the axe-cucumber step
+# Requiring 'axe-cucumber-steps' makes all of the axe-cucumber step
 # definitions available to be used directly in your cucumber features.
-require "axe-core-cucumber-step-definitions"
+require "axe-cucumber-steps"
 # configure `AxeCapybara`
 @page = AxeCapybara.configure(:firefox) do |c|
 end
