+++ packages/axe-core-cucumber/e2e/watir/features/support/env.rb
@@ -2,9 +2,9 @@
 require "watir"
 # require driver of choice
 require "axe-watir"
-# Requiring 'axe/cucumber/step_definitions' makes all of the axe-cucumber step
+# Requiring 'axe-cucumber-steps' makes all of the axe-cucumber step
 # definitions available to be used directly in your cucumber features.
-require "axe-core-cucumber-step-definitions"
+require "axe-cucumber-steps"
 
 Before do
   # configure AxeWatir
