+++ packages/axe-core-cucumber/lib/axe-cucumber-steps.rb
@@ -1,3 +1,3 @@
-require_relative "./axe-core-cucumber"
+require_relative "./axe-cucumber"
 
 Then AxeCucumber::Step::REGEX, :assert_accessibility, on: -> { AxeCucumber::Step.create_for self }
