+++ packages/axe-core-cucumber/spec/step_regex_spec.rb
@@ -1,5 +1,5 @@
 require "spec_helper"
-require_relative "../lib/axe-core-cucumber"
+require_relative "../lib/axe-cucumber"
 
 module CustomMatchers
   class Match < RSpec::Matchers::BuiltIn::Match
