+++ packages/axe-core-watir/spec/axe-watir_spec.rb
@@ -10,7 +10,7 @@ describe AxeWatir do
       end
 
       expect(driver).not_to be_nil
-      driver.page.goto "https://google.com" # can navigate
+      driver.page.goto "https://deque.com" # can navigate
       expect(driver).to respond_to :skip_iframes # can config
       expect(driver).to respond_to :jslib
       expect(driver.jslib).to include("axe.run=") # has axe injected
